requirements.txt
pytz
requests
requests[socks]
pyserial
smspdudecoder